+++ src/elements/robots/RobotAPattes.java
@@ -9,7 +9,7 @@ import elements.WrongCaseNatureException;
 public class RobotAPattes extends Robot {
 
     private final static int capacite = Integer.MAX_VALUE;
-    private final static int vitesse = 30;
+    private final static double vitesse = 30;
     private final static double vitesseIntervention = 10;
 
     public RobotAPattes(Carte carte) {
@@ -53,7 +53,7 @@ public class RobotAPattes extends Robot {
     }
 
     @Override
-    public int getVitesse(NatureTerrain T) {
+    public double getVitesse(NatureTerrain T) {
         if (T == NatureTerrain.EAU) {
             return 0;
         }
